token: add tests, fix totalSupply bug

+++ src/erc20.sol
@@ -2,15 +2,15 @@
 pragma solidity ^0.8.4;
 
 interface ERC20 {
-    function balanceOf(address usr) external view returns (uint256);
-    function totalSupply() external view returns (uint256);
+    function balanceOf(address usr) external view returns (uint);
+    function totalSupply() external view returns (uint);
 
-    function allowance(address src, address dst) external view returns (uint256);
-    function approve(address dst, uint256 amt) external returns (bool);
+    function allowance(address src, address dst) external view returns (uint);
+    function approve(address dst, uint amt) external returns (bool);
 
-    function transfer(address dst, uint256 amt) external returns (bool);
-    function transferFrom(address src, address dst, uint256 amt) external returns (bool);
+    function transfer(address dst, uint amt) external returns (bool);
+    function transferFrom(address src, address dst, uint amt) external returns (bool);
 
-    event Transfer(address indexed src, address indexed dst, uint256 amt);
-    event Approval(address indexed src, address indexed dst, uint256 amt);
+    event Transfer(address indexed src, address indexed dst, uint amt);
+    event Approval(address indexed src, address indexed dst, uint amt);
 }
